ci(enforce): wire the 3 remaining local-hook-only gates into CI (Wave 1 / #195)
Closes the local-vs-CI divergence for the last three Wave-1 gates that ran
only in the bypassable pre-commit hook (check_hil_alive_policy, the fourth,
landed in 5a2c67a5).

check_assert_casts -- clean the tree + wire tests/*.c
  * Stripped 1485 redundant (int)/(uint32_t) casts from TEST_ASSERT_EQ
    arguments across 104 test files. The macro widens both args to int64_t,
    so an outer cast is redundant and latently buggy (an (int) cast on a
    uint32_t enum truncates before the widening).
  * Five casts were load-bearing, not redundant, and are hoisted into typed
    locals -- value preserved, assertion cast-free:
      - test_ra_usb_xfer_cov.c    (uint16_t)~x complement truncation
      - test_ra_fs_fat_name_cov.c signed char 0xE5 -> unsigned byte (229)
      - test_app_dtc_transfer_demo.c (x3) 64-bit host pointer -> 32-bit
        SAR/DAR/vector register width (HUM Ch 18.2)
  * Wired check_assert_casts.py tests/*.c into the pre-commit-checks CI job.

check_annotations -- make libclang-robust + wire strict
  * cindex.CursorKind.SECTION_ATTR is absent from the libclang 18.1.x wheels;
    merely accessing it raised AttributeError and aborted the TU walk, silently
    dropping every TU that defined an annotated function in place (ra_widget_*,
    ra_book, ra_rabook_*). Resolve the kind once via getattr(...) with a None
    fallback so it degrades to "no section detected" -- all 1282 TUs now walk.
  * The tree clears --check (0 violations), so the pre-commit hook and a new
    CI step run it strict. CI installs a pinned libclang wheel into a venv;
    local keeps WARN_ONLY_MODE=True so a dev without python-clang skips
    gracefully instead of being blocked.

stack_usage_check --strict -- fix discovery, shrink frames, wire after build
  * find_su_files only walked examples/<tier>/<app>/build* (2 levels) while
    every EK firmware app nests 3-4 levels deep, so the aggregator had never
    actually scanned them -- it found 0 .su and passed trivially. Recurse to
    any depth and derive the app name from the path component before build*/.
  * The now-real gate surfaced 4 first-party frames over the 2048 B budget,
    all large per-call DFS/cascade scratch stacks. Moved to module-static
    storage (the gate's own remedy); each walk is iterative (no recursion),
    single-threaded, and has a single call site, so the shared buffer never
    overlaps:
      priv_reflow_xml_walk      2168 -> 32   ra_reflow_tokenize.c
      ra_book_walk_to_xhtml     4160 -> 72   ra_book_xhtml.c
      ra_book_walk_text         2104 -> 56   ra_book_xhtml.c
      ra_book_walk_text_paged   2112 -> 64   ra_book_paged.c
  * Wired python3 stack_usage_check.py --strict into the build-cross CI job,
    after build_all_examples.sh where the .su files exist. --strict is green:
    0 first-party frames > 2048 B, 0 dynamic (VLA/alloca) frames, 0 critical-
    module breaches. The remaining soft violations are all third-party SOUP
    (miniz/stb), which --strict exempts. Doc updated (docs/STACK_USAGE.md).

Validated on the dev box (CI-parity host toolchain): 430/430 host tests pass,
format_code.sh --check clean, all three gates green. Stack frames measured
against the arm-gcc cross build (margin is large -- the next first-party frame
is 1704 B).

(cherry picked from commit f5ad788eac21e69830c0ced378d5168b0aceb282)

+++ tests/test_ra_cgc.c
@@ -244,13 +244,13 @@ static void test_init_pfb_threshold(void)
   *ra_sys_oscsf()    = (uint8_t)k_ra_cgc_test_all_oscsf;
   *ra_mrms_mrcfreq() = (uint32_t)k_ra_cgc_test_mrcfreq_fast;
   TEST_ASSERT_EQ(k_ra_ok, ra_cgc_init());
-  TEST_ASSERT_EQ((uint32_t)k_ra_mrcpfb_enable, *ra_mrms_mrcpfb());
+  TEST_ASSERT_EQ(k_ra_mrcpfb_enable, *ra_mrms_mrcpfb());
 
   /* V2: MRCFREQ reads 0 -> below threshold -> PFB stays disabled. */
   cgc_test_reset();
   *ra_sys_oscsf() = (uint8_t)k_ra_cgc_test_all_oscsf;
   TEST_ASSERT_EQ(k_ra_ok, ra_cgc_init());
-  TEST_ASSERT_EQ((uint32_t)k_ra_mrcpfb_disable, *ra_mrms_mrcpfb());
+  TEST_ASSERT_EQ(k_ra_mrcpfb_disable, *ra_mrms_mrcpfb());
   TEST_END("cgc init PFB threshold arms");
 }
